+++ crates/notedeck_columns/src/ui/configure_deck.rs
@@ -1,10 +1,9 @@
 use crate::{app_style::deck_icon_font_sized, deck_state::DeckState};
 use egui::{vec2, Button, Color32, Label, RichText, Stroke, Ui, Widget};
 use notedeck::{NamedFontFamily, NotedeckTextStyle};
-use notedeck_ui::colors::PINK;
-
-use super::{
+use notedeck_ui::{
     anim::{AnimationHelper, ICON_EXPANSION_MULTIPLE},
+    colors::PINK,
     padding,
 };
