+++ crates/notedeck_ui/src/note/options.rs
@@ -1,4 +1,4 @@
-use crate::ui::ProfilePic;
+use crate::ProfilePic;
 use bitflags::bitflags;
 
 bitflags! {
