+++ crates/notedeck_ui/src/note/reply_description.rs
@@ -1,12 +1,10 @@
-use crate::{
-    actionbar::NoteAction,
-    ui::{self},
-};
 use egui::{Label, RichText, Sense};
-use enostr::KeypairUnowned;
 use nostrdb::{Note, NoteReply, Transaction};
 
-use super::{contents::NoteContext, NoteOptions};
+use super::NoteOptions;
+use crate::{note::NoteView, Mention};
+use enostr::KeypairUnowned;
+use notedeck::{NoteAction, NoteContext};
 
 #[must_use = "Please handle the resulting note action"]
 #[profiling::function]
@@ -41,7 +39,7 @@ pub fn reply_desc(
             if r.hovered() {
                 r.on_hover_ui_at_pointer(|ui| {
                     ui.set_max_width(400.0);
-                    ui::NoteView::new(note_context, cur_acc, note, note_options)
+                    NoteView::new(note_context, cur_acc, note, note_options)
                         .actionbar(false)
                         .wide(true)
                         .show(ui);
@@ -62,7 +60,7 @@ pub fn reply_desc(
 
     if note_reply.is_reply_to_root() {
         // We're replying to the root, let's show this
-        let action = ui::Mention::new(
+        let action = Mention::new(
             note_context.ndb,
             note_context.img_cache,
             txn,
@@ -86,7 +84,7 @@ pub fn reply_desc(
         if let Ok(root_note) = note_context.ndb.get_note_by_id(txn, root.id) {
             if root_note.pubkey() == reply_note.pubkey() {
                 // simply "replying to bob's note" when replying to bob in his thread
-                let action = ui::Mention::new(
+                let action = Mention::new(
                     note_context.ndb,
                     note_context.img_cache,
                     txn,
@@ -109,7 +107,7 @@ pub fn reply_desc(
             } else {
                 // replying to bob in alice's thread
 
-                let action = ui::Mention::new(
+                let action = Mention::new(
                     note_context.ndb,
                     note_context.img_cache,
                     txn,
@@ -134,7 +132,7 @@ pub fn reply_desc(
                     Label::new(RichText::new("in").size(size).color(color)).selectable(selectable),
                 );
 
-                let action = ui::Mention::new(
+                let action = Mention::new(
                     note_context.ndb,
                     note_context.img_cache,
                     txn,
@@ -156,7 +154,7 @@ pub fn reply_desc(
                 note_link(ui, note_context, "thread", &root_note);
             }
         } else {
-            let action = ui::Mention::new(
+            let action = Mention::new(
                 note_context.ndb,
                 note_context.img_cache,
                 txn,
